[#210] Workaround for 'rust-lang/cargo#6313'
