+++ iceoryx2-ffi/ffi/src/tests/node_tests.rs
@@ -10,16 +10,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod common;
-
 #[generic_tests::define]
 mod node {
-
-    use crate::common::*;
-
-    use iceoryx2::prelude::*;
-    use iceoryx2_bb_testing::assert_that;
-    use iceoryx2_ffi::*;
+    use crate::tests::*;
 
     use core::{slice, str};
